restructure flux code
